add data quality verifier and letter-based class system
- Add data/verify_data.py: validates landmarks.pickle with per-class
  stats, bad sample detection, imbalance warnings, and --fix flag
- Switch class folders from numbered (0, 1, 2) to letter-named (a, b, c)
- Add CLASSES list to config.py, derive NUM_CLASSES from it
- Update collect_images.py to use letter labels in prompts and folders
- Update extract_landmarks.py to support string labels
- Handle empty/corrupt pickle files gracefully in verifier
- Add .gitkeep to data/raw/ so folder structure is preserved on GitHub

+++ data/extract_landmarks.py
@@ -82,7 +82,7 @@ def extract_landmarks_batch(
     )
 
     for class_folder in class_folders:
-        class_label = int(class_folder.name)
+        class_label = class_folder.name
         logging.info(f"processing class {class_label}...")
 
         images = list(class_folder.glob("*.jpg")) + \
@@ -126,7 +126,7 @@ def extract_landmarks_batch(
     logging.info("converting to numpy arrays...")
 
     all_data = np.array(all_data, dtype=np.float32)
-    all_labels = np.array(all_labels, dtype=np.int32)
+    all_labels = np.array(all_labels)
 
     output_dict = {
         "data": all_data,
